// src/app/interfaces/background.ts
import { backgroundColor, backgroundStyle, patternColor, patternSize } from './canvas-settings';

interface BackgroundInterface {
    backgroundStyle: backgroundStyle;
    backgroundColor: backgroundColor;
    patternSize: patternSize;
    patternColor: patternColor;
    // opacity: number;
}
export { BackgroundInterface };